+++ src/modules/menu.ts
@@ -1,11 +1,11 @@
-import { Menu, dialog, app, BrowserWindow, shell } from 'electron';
+import { Menu, dialog, app, BrowserWindow } from 'electron';
 import { getAppVersion, getCurrentUrl, getEnv, isDev } from './config';
 import { createMainWindow, getMainWindow } from './windowManager';
 import { reloadMainWindow } from './reloadUtils';
 import { safeClosewhiteboardWindow } from './whiteboard-window';
 import { safeCloseScreenShareWindow } from './screenShareWindow';
 import { safeCloseStreamWindow } from './streamWindow';
-import { setActiveProfile } from '@/utils/profileUtils';
+import { clearActiveProfileStorage } from '@/utils/profileUtils';
 
 /**
  * Opens a Chrome internal URL in a new window
@@ -121,8 +121,8 @@ function createMenu(): void {
         {
           label: 'Quit',
           accelerator: process.platform === 'darwin' ? 'Cmd+Q' : 'Ctrl+Q',
-          click: () => {
-            setActiveProfile(null);
+          click: async () => {
+            await clearActiveProfileStorage();
             app.quit();
           },
         },
